refactor: inline Permissions concern into Member
Permissions was a single-model concern only included in Member, making it
a code-slicing concern rather than a shared behavioral extraction. Inlined
its rolify call and 4 instance methods (organiser?, admin_or_organiser?,
monthlies_organiser?, organised_chapters) directly into Member.

- Removed app/models/concerns/permissions.rb
- Updated spec to reference Member instead of Permissions module
