+++ app/models/member.rb
@@ -1,6 +1,7 @@
 class Member < ApplicationRecord
-  include Permissions
   include DigestHelper
+
+  rolify role_cname: 'Permission', role_table_name: :permission, role_join_table_name: :members_permissions
 
   enum :how_you_found_us, {
     from_a_friend: 0,
@@ -63,6 +64,22 @@ class Member < ApplicationRecord
   acts_as_taggable_on :skills
 
   attr_accessor :attendance, :newsletter
+
+  def organiser?
+    organised_chapters.present?
+  end
+
+  def admin_or_organiser?
+    has_role?(:admin) || organiser?
+  end
+
+  def monthlies_organiser?
+    Meeting.with_role(:organiser, self).present?
+  end
+
+  def organised_chapters
+    Chapter.with_role(:organiser, self)
+  end
 
   def manager?
     is_admin? || has_role?(:organiser, :any)
